requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bleach==2.1.4
boto==2.49.0
boto3==1.7.75
botocore==1.10.75
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
cryptography==2.3
decorator==4.3.0
diff-match-patch==20121119
Django==1.11.15
django-anymail==5.0
-e git+https://github.com/django-import-export/django-import-export.git@54232db4a511cc5f0b085c6bb192c3619f27e452#egg=django_import_export
django-rest-framework==0.1.0
django-storages==1.7
djangorestframework==3.8.2
djangorestframework-datatables==0.4.0
djrill==2.1.0
dlib==19.15.0
docutils==0.14
dropbox==9.0.0
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
face-recognition==1.2.2
face-recognition-models==0.3.0
functools32==3.2.3.post2
futures==3.2.0
html5lib==1.0.1
idna==2.7
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jdcal==1.4
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.3.0
MarkupSafe==1.0
mistune==0.8.3

nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.0
odfpy==1.3.6
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
openpyxl==2.5.6
pandocfilters==1.4.2
pathlib==1.0.1
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.2.0

python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
s3transfer==0.1.13
scandir==1.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
soupsieve==1.7.2
tablib==0.12.1
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.0
xlrd==1.1.0
xlwt==1.3.0
